Applied patch from http://citylight.thinkbay.net/Disable_sharing_in_rtorrent
Disable sharing in rtorrent
Fri, 07/25/2008 - 16:06 — lukasz

If, for some reason, you need to totally disable sharing in rtorrent (and every other application using libtorrent) and look as if you always had 0% of the files you are downloading, download the source of libtorrent and patch it with the attached diff.

I take no responsibility for this patch, use it at your own risk.

Moreover, use it only if, for whatever reason, you really really need to disable sharing. If you are short on bandwidth, you can always set upload throttling to, say, 1 kBps instead of patching and recompiling everything.
It is much simpler this way and you can still help the swarm.

Remember: people will STILL see you are connected to the tracker and if a client is using Superseeding, you will NOT GET ANY OTHER CHUNK of the file until you upload what you just downloaded to someone else! (which won't occur since you just disabled sharing).

Consider yourself warned.